requirements.txt
PyYAML
numpy
matplotlib
seaborn
pytest